Modify the monolithic approach to be notebook compatible with the modular approach. This means instead of passing a STAC catalog filename the directory is passed with the expectation of there being a file name catalog.json inside it.
Also restructure some of the CWL and Descriptor classes.
