+++ app_pack_generator/__init__.py
@@ -1,6 +1,7 @@
 from .git import GitManager, GitRepoError
 from .docker import DockerUtil
 from .application import ApplicationNotebook
-from .process import ProcessCWL, BundledDataStagingCWL, Descriptor
+from .cwl import ProcessCWL, DataStagingCWL
+from .descriptor import Descriptor
 
 from .version import __version__
