+++ app_pack_generator/cwl.py
@@ -1,6 +1,5 @@
 import os
 import re
-import json
 import yaml
 import logging
 
@@ -15,7 +14,7 @@ def write_cwl_file(fname, target):
         f.write("#!/usr/bin/env cwl-runner\n")
         yaml.dump(target, f, default_flow_style=False)
 
-class ProcessError(Exception):
+class CWLError(Exception):
     pass
 
 class BaseCWL(object):
@@ -71,7 +70,7 @@ class ProcessCWL(BaseCWL):
                 'default': param.default,
             }
 
-    def _insert_input_params(self, input_filename=None):
+    def _insert_input_params(self):
         "Connects the special 'input' CWL parameter to the papermill parameter for recieving the stage-in directory location"
 
         input_dict = self.process_cwl['inputs']
@@ -86,16 +85,10 @@ class ProcessCWL(BaseCWL):
         if self.app.stage_in_param is not None:
             input_dict['input'] = 'Directory'
 
-            if input_filename is None:
-                input_filename = ""
-            elif len(input_filename) != 0 and input_filename[0] != "/":
-                # Make sure if input_filename has a path seperator if not already given
-                input_filename = f"/{input_filename}"
-
             self.process_cwl['arguments'] = self.process_cwl.get('arguments', [])
             self.process_cwl['arguments'] += [
                 '-p', self.app.stage_in_param.name,
-                f'$(inputs.input.path){input_filename}'
+                f'$(inputs.input.path)'
             ]
     
     def _insert_output_params(self):
@@ -109,7 +102,7 @@ class ProcessCWL(BaseCWL):
             stage_out_process_dir = self.process_cwl['outputs']['output']['outputBinding']['glob']
 
             if not re.search('runtime.outdir', stage_out_process_dir):
-                raise ProcessError(f"The process CWL template outputs/output path needs to contain $(runtime.outdir) in the path")
+                raise CWLError(f"The process CWL template outputs/output path needs to contain $(runtime.outdir) in the path")
 
             self.process_cwl['arguments'] = self.process_cwl.get('arguments', [])
             self.process_cwl['arguments'] += [
@@ -143,10 +136,10 @@ class ProcessCWL(BaseCWL):
         return fname
 
 
-class BundledDataStagingCWL(ProcessCWL):
+class DataStagingCWL(BaseCWL):
 
     def __init__(self, application, **kwargs):
-
+        
         super().__init__(application, **kwargs)
 
         # Template CWL and descriptor files
@@ -154,15 +147,14 @@ class BundledDataStagingCWL(ProcessCWL):
         self.stage_in_cwl = self._read_template( os.path.join(self.template_dir, 'stage_in.cwl'))
         self.stage_out_cwl = self._read_template( os.path.join(self.template_dir, 'stage_out.cwl'))
 
-
     def generate_all(self, outdir, **kwargs):
         """Calls all of the application CWL generators as well as the application descriptor generator.
 
         Returns the list of all files generated by this function (abs. path).
         """
 
-        generated_files = super().generate_all(outdir, **kwargs)
-        
+        generated_files = []
+
         generated_files.append(self.generate_workflow_cwl(outdir))
         generated_files.append(self.generate_stage_in_cwl(outdir))
         generated_files.append(self.generate_stage_out_cwl(outdir))
@@ -237,92 +229,4 @@ class BundledDataStagingCWL(ProcessCWL):
         # Generate the outputs CWL as-is, no need for modifications
         fname = os.path.join(outdir, 'stage_out.cwl')
         write_cwl_file(fname, self.stage_out_cwl)
-        return fname
-
-    def _insert_input_params(self):
-        "Overrides ProcessCWL implementation to use old behavior of specifying a file instead of directory for stage-in"
-
-        # We construct the stage in argument to point to the download directory 
-        # with the filename mentioned in the stage_in CWL file. This replicates the old
-        # monolithic application generation behavior.
-        #
-        # If we change the filename name in the stage in CWL template it is reflected here.
-        input_dict = self.process_cwl['inputs']
-
-        if self.app.stage_in_param is not None:
-            stage_in_collection_filename = self.stage_in_cwl['outputs']['stage_in_collection_file']['outputBinding']['glob']
-
-            super()._insert_input_params(stage_in_collection_filename)
-
-
-class Descriptor(object):
-
-    def __init__(self, application, repo, templatedir=os.path.join(LOCAL_PATH, 'templates')):
-
-        self.app = application
-        self.repo = repo
-
-        self.descriptor = self._read_template(os.path.join(templatedir, 'app_desc.json'))
-
-    def _read_template(self, desc_fname):
-        """Loads the template application descriptor and should validate it to
-        ensure an exception is not thrown.
-        """
-        with open(desc_fname, 'r') as f:
-            return json.load(f)
-
-    def generate_descriptor(self, outdir, dockerurl):
-        """Generates the application descriptor JSON.
-
-        Returns the absolute  path of the file generated by this function.
-        """
-        if not os.path.isdir(outdir):
-            os.makedirs(outdir)
-        deposit_url = 'https://raw.githubusercontent.com/jplzhan/artifact-deposit-repo'
-        tag = self.repo.name
-        proc_dict = self.descriptor['processDescription']['process']
-
-        if self.repo.owner is not None:
-            proc_dict['id'] = self.repo.owner + '.' + \
-                self.repo.name + '.' + self.repo.commit_identifier
-        else:
-            proc_dict['id'] = self.repo.name + '.' + self.repo.commit_identifier
- 
-        proc_dict['title'] = self.repo.commit_message
-        proc_dict['owsContext']['offering']['content']['href'] = deposit_url + \
-            '/main/' + tag + '/workflow.cwl'
-
-        proc_dict['inputs'] = []
-        for param in self.app.arguments:
-            key_type = param.cwl_type
-            proc_dict['inputs'].append({
-                'id': param.name, 
-                'title': 'Automatically detected using papermill.',
-                'literalDataDomains': [{'dataType': {'name': key_type}}],
-            })
-
-        if self.app.stage_in_param is not None:
-            proc_dict['inputs'].append({
-                'id': self.app.stage_in_param.name,
-                'title': 'Stage-in input specified for URL-to-PATH conversion.',
-                'literalDataDomains': [{'dataType': {'name': 'stage_in'}}],
-            })
-
-        proc_dict['outputs'] = []
-        if self.app.stage_out_param is not None:
-            proc_dict['outputs'].append({
-                'id': self.app.stage_out_param.name,
-                'title': 'Automatically detected from .ipynb parsing.',
-                'output': {
-                    'formats': [
-                        {'mimeType': 'text/*', 'default': True}
-                    ]
-                }
-            })
-
-        self.descriptor['executionUnit'][0]['href'] = 'docker://' + dockerurl
-
-        fname = os.path.join(outdir, 'applicationDescriptor.json')
-        with open(fname, 'w', encoding='utf-8') as f:
-            json.dump(self.descriptor, f, ensure_ascii=False, indent=4)
         return fname
